x_pub_article/requirements.txt
playwright>=1.41.0
Pillow>=10.0.0
numpy>=1.24.0
